+++ tests/fixture/tmpnet/defaults.go
@@ -51,7 +51,9 @@ func DefaultTmpnetFlags() FlagsMap {
 // Flags suggested for e2e testing
 func DefaultE2EFlags() FlagsMap {
 	return FlagsMap{
-		config.ProposerVMUseCurrentHeightKey: "true",
+		config.SystemTrackerRequiredAvailableDiskSpacePercentageKey: "1",
+		config.SystemTrackerWarningAvailableDiskSpacePercentageKey:  "1",
+		config.ProposerVMUseCurrentHeightKey:                        "true",
 		// Reducing this from the 1s default speeds up tx acceptance
 		config.ProposerVMMinBlockDelayKey: "0s",
 		config.LogLevelKey:                logging.Debug.String(),